Refactor WelcomeViewModel environment initialization
